requirements.txt
flask==3.0.3
flask_sqlalchemy==3.1.1
sqlalchemy==2.0.36
networkx==3.3
numpy==2.1.2
pymongo==4.9.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
asgiref==3.8.1
uvicorn==0.30.6